extract ReportNavigateAwayHandler from ReportScreen
